back/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
PyMySQL==1.1.2
python-dotenv==1.1.1
pydantic-settings==2.10.1
email-validator==2.2.0
google-auth==2.40.3
python-multipart==0.0.20
Pillow==11.3.0
